% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSimians.R
\docType{data}
\name{DataSimians}
\alias{DataSimians}
\title{Example dataset}
\usage{
data(DataSimians)
}
\description{
The output of Procrustes superimposition as performed by the
  function \code{\link[Morpho]{procSym}} on 9 simians faces (\code{$pca}),
  the phylogenetic tree for such species (\code{$tree}), and the 3d surfaces
  along with landmark configurations for \emph{Pan troglodytes}
  (\code{$sur_pan} and \code{$ldm_pan}) and \emph{Alouatta caraya}
  (\code{$sur_alo} and \code{$ldm_alo}).
}
\author{
Pasquale Raia, Silvia Castiglione, Carmela Serio, Alessandro
  Mondanaro, Marina Melchionna, Antonio Profico
}
\keyword{RRmorph}
