% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLegend.R
\name{plotLegend}
\alias{plotLegend}
\title{Plot legend for colored mesh.}
\usage{
plotLegend(mesh,values,main)
}
\arguments{
\item{mesh}{a \code{mesh3d} object}

\item{values}{a vector of continuous values associated to individual vertices
of the \code{mesh}.}

\item{main}{plot title.}
}
\value{
A plot of the color sequence associated to \code{values} on the \code{mesh}.
}
\description{
Assuming a \code{mesh} is colored according to a vector of \code{values},
  the function takes the color sequence from the \code{mesh} and plots it
  associated to \code{values}.
}
\examples{
  \dontrun{
  require(rgl)
  data(DataSimians)
  pca<-DataSimians$pca
  ldm_pan<-DataSimians$ldm_pan
  sur_pan<-DataSimians$sur_pan

  rec_pan<- Rvcg::vcgBallPivoting(pca$mshape, radius = 0)
  rec_pan$vb[1:3,]<-t(ldm_pan)
  val<-rnorm(ncol(rec_pan$vb))

  interp<-interpolMesh(sur = rec_pan,refsur = sur_pan,refmat = ldm_pan,
                       values = val,element ="vertices",k = 4)

  colmesh<-col2mesh(mesh = sur_pan,values = interp,pal = heat.colors(5))
  plotLegend(mesh = colmesh,values = interp, main = "Pan troglodytes")
  open3d()
  shade3d(colmesh,specular="black")
  }
}
\author{
Marina Melchionna, Silvia Castiglione
}
