% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.map.R
\name{rate.map}
\alias{rate.map}
\title{Mapping rate and direction of phenotypic change on 3D surfaces.}
\usage{
rate.map(x, RR, scores, pcs, mshape, mshape_sur=NULL,refsur=NULL,
  refmat=NULL, k=4,out.rem = TRUE, plot=TRUE, pal=NULL,
  NAcol="gray90",from=NULL, to=NULL,show.names=TRUE)
}
\arguments{
\item{x}{the species/nodes to be compared; it can be a single species, or a
vector containing two species or a species and any of its parental nodes.}

\item{RR}{an object generated by using the \code{\link{RRphylo}} function}

\item{scores}{RW or PC scores.}

\item{pcs}{RW (or PC) vectors (eigenvectors of the covariance matrix)
returned by RWA/PCA.}

\item{mshape}{the consensus configuration.}

\item{mshape_sur}{a \code{mesh3d} object used as a reference for mesh
reconstruction. The vertices of \code{mshape_sur} must be the consensus
configuration. If \code{NULL}, it is automatically generated by applying
\code{\link[Rvcg]{vcgBallPivoting}} on \code{mshape}.}

\item{refsur}{a list of \code{mesh3d} to be provided for all the species in
\code{x}.}

\item{refmat}{a list of landmark sets to be provided for all the species in
\code{x}.}

\item{k}{the argument \code{k} passed to \code{\link{interpolMesh}}.}

\item{out.rem}{logical: if \code{TRUE} mesh triangles with outlying area
difference are removed.}

\item{plot}{logical indicating if the 3d plot must be shown.}

\item{pal}{a vector of color to be passed to
\code{\link[grDevices]{colorRampPalette}}.}

\item{NAcol}{the argument \code{NAcol} passed to \code{\link{col2mesh}}.}

\item{from, to}{lower and upper limits to be associated to the ends of
\code{pal}.}

\item{show.names}{logical: if \code{TRUE}, the names of the species are
displayed in the 3d plot.}
}
\value{
The function returns a list including:
  \itemize{\item\strong{$selected} a list of PCs axes selected for higher
  evolutionary rates for each species. \item\strong{$surfaces} a list of
  reconstructed colored surfaces of the given species and of the most recent
  common ancestor.\item\strong{lmks} . \item\strong{differences} }


  selected=rates.list,surfaces=c(mrca=list(sur.ref),meshes),lmks=refmat,differences=values
}
\description{
Given vectors of RW (or PC) scores, the function selects the RW
  (PC) axes linked to highest (and lowest) evolutionary rate values and
  reconstructs the morphology weighted on them. In this way, \code{rate.map}
  shows where and how the phenotype changed the most between any pair of
  taxa.
}
\details{
After selecting the RW (PC) axes, \code{rate.map} automatically builds a
  3D mesh on the mean shape calculated from the Relative Warp Analysis (RWA)
  or Principal Component Analysis (PCA) (\cite{Schlager 2017}) by applying
  the function \code{\link[Rvcg]{vcgBallPivoting}} (\pkg{Rvcg}). Then, it
  compares the area differences between corresponding triangles of the 3D
  surfaces reconstructed for the species and surface of the mrca. Finally,
  \code{rate.map} returns a 3D plot showing such comparisons displayed on
  shape of the mrca used as the reference.The color gradient goes from blue
  to red, where blue areas represent expansion of the mesh, while the red
  areas represent contraction of the mesh triangles. If a list \code{refsur}
  (and \code{refmat}) is provided, convergence is plotted onto them (see
  \code{\link{interpolMesh}} for further details). In the calculation of the
  differences of areas we supply the possibility to find and remove outliers
  from the vectors of areas calculated on the reference and target surfaces.
  We suggest considering this possibility if the mesh may contain degenerate
  facets.
}
\examples{
  \dontrun{
  data(DataSimians)
  pca<-DataSimians$pca
  tree<-DataSimians$tree
  dato<-pca$PCscores
  cc<- 2/parallel::detectCores()

  RR<-RRphylo::RRphylo(tree,dato,clus=cc)

  # plotting on reconstructed surfaces
  Rmap1<-rate.map(x=c("Pan_troglodytes","Alouatta_caraya"),RR=RR, scores=dato,
                 pcs=pca$PCs, mshape=pca$mshape)

  # plotting on real surfaces
  Rmap2<-rate.map(x=c("Pan_troglodytes","Alouatta_caraya"),RR=RR, scores=dato,
                 pcs=pca$PCs, mshape=pca$mshape,
                 refsur=list("Pan_troglodytes"=sur_pan,"Alouatta_caraya"=sur_alo),
                 refmat=list("Pan_troglodytes"=ldm_pan,"Alouatta_caraya"=ldm_alo))



  }
}
\references{
Schlager, S. (2017). \emph{Morpho and Rvcg-Shape Analysis in R:
  R-Packages for geometric morphometrics, shape analysis and surface
  manipulations.} In: Statistical shape and deformation analysis. Academic
  Press. Castiglione, S., Melchionna, M., Profico, A., Sansalone, G.,
  Modafferi, M., Mondanaro, A., Wroe, S., Piras, P., & Raia, P. (2021). Human
  face-off: a new method for mapping evolutionary rates on three-dimensional
  digital models. Palaeontology. doi:10.1111/pala.12582
}
\seealso{
\href{../doc/RRphylo.html}{\code{RRphylo} vignette} ;
  \code{\link[Morpho]{relWarps}} ; \code{\link[Morpho]{procSym}}
}
\author{
Marina Melchionna, Antonio Profico, Silvia Castiglione, Gabriele
  Sansalone, Pasquale Raia
}
