% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapeDiff.R
\name{shapeDiff}
\alias{shapeDiff}
\title{Shape difference between 3d meshes}
\usage{
shapeDiff(x,relw,refsur,refmat,mshape_sur = NULL,
  pal=NULL,NAcol="gray90",show.names=TRUE)
}
\arguments{
\item{x}{a vector of specimens pair.}

\item{relw}{the result of a relative warp analysis. Classes \code{relwarps} and
\code{nosymproc} are both accepted.}

\item{refsur}{a list of two \code{mesh3d} objects to be provided in the same order as \code{x}.}

\item{refmat}{a list of two landmark sets related to \code{refsur} to be
provided in the same order as \code{x}.}

\item{mshape_sur}{a \code{mesh3d} object used as a reference for mesh reconstruction.
The vertices of \code{mshape_sur} must be the consensus configuration. If \code{NULL},
it is automatically generated by applying \code{\link[Rvcg]{vcgBallPivoting}}
on the consensus configuration derived from \code{relw}.}

\item{pal}{a vector of color to be passed to \code{\link[grDevices]{colorRampPalette}}.}

\item{NAcol}{the color associated to \code{refsur} vertices falling outside the
range of \code{refmat} (not involved in interpolation).}

\item{show.names}{logical: if \code{TRUE}, the names of the specimens as in \code{x}
are displayed in the 3d plot.}
}
\value{
Two \code{mesh3d} objects colored according to shape differences.
  Additionally, the function returns 3d plots of the meshes.
}
\description{
The function reconstructs two specimens' meshes (\code{x})
  by using their superimposed configurations (from within \code{relw}),
  calculates the shape difference between them, and plots such differences
  on provided meshes (\code{refsur}).
}
\examples{
  \dontrun{
  require(rgl)
  data(DataSimians)
  pca<-DataSimians$pca
  ldm_pan<-DataSimians$ldm_pan
  sur_pan<-DataSimians$sur_pan
  ldm_alo<-DataSimians$ldm_alo
  sur_alo<-DataSimians$sur_alo

  diffs<-shapeDiff(x=c("Pan_troglodytes","Alouatta_caraya"),
                   relw = pca,refsur = list(sur_pan,sur_alo),
                   refmat = list(ldm_pan,ldm_alo))
  }
}
\author{
Marina Melchionna, Silvia Castiglione
}
