% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri2verts.R
\name{tri2verts}
\alias{tri2verts}
\title{Triangles to vertices}
\usage{
tri2verts(mesh,values)
}
\arguments{
\item{mesh}{a \code{mesh3d} object.}

\item{values}{a vector of continuous values associated to individual triangles
of the \code{mesh}.}
}
\value{
A vector of continuous values associated to individual vertices of
  the \code{mesh}.
}
\description{
The function transfers values associated to triangles of a
  \code{mesh3d} object to its vertices.
}
\examples{
  \dontrun{
  require(rgl)
  data(DataSimians)
  pca<-DataSimians$pca
  ldm_pan<-DataSimians$ldm_pan
  sur_pan<-DataSimians$sur_pan

  rec_pan<- Rvcg::vcgBallPivoting(pca$mshape, radius = 0)
  rec_pan$vb[1:3,]<-t(ldm_pan)
  val<-rnorm(ncol(rec_pan$vb))

  vertval<-tri2verts(rec_pan,val)
  }
}
\author{
Marina Melchionna, Silvia Castiglione
}
